+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Changelog
   simulated in V-REP.
 - Added copying and pasting object via interface to a generic scene object
   simulated in V-REP.
+- Added removing object from scene via interface to a generic scene object
+  simulated in V-REP.
 - Added removing model from scene via interface to a generic model simulated in
   V-REP.